Move game logic into a new component
