+++ src/App.css
@@ -5,9 +5,3 @@
   text-align: center;
   font-size: x-large;
 }
-
-.game-board{
-  display: grid;
-  gap: 1ch;
-  max-width: max-content;
-}
