+++ src/components/Game/Game.css
@@ -1,0 +1,6 @@
+
+.game-board{
+  display: grid;
+  gap: 1ch;
+  max-width: max-content;
+}
